requirements.txt
langchain==0.1.16
langchain_core==0.1.43
python-decouple==3.8
Requests==2.31.0
streamlit==1.33.0
openai